requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
send2trash>=1.8.0
